# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/functions": {
            "get": {
                "description": "Retrieves a list of all registered functions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "functions"
                ],
                "summary": "List all functions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/functions.Function"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Uploads a Python file, creates a new FaaS function container, and returns its details.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "functions"
                ],
                "summary": "Add a new function",
                "parameters": [
                    {
                        "type": "file",
                        "description": "The Python file containing the function handler",
                        "name": "python_file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The name of the function to execute (e.g., 'handle')",
                        "name": "function_name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/functions.Function"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/functions/{functionID}": {
            "delete": {
                "description": "Stops the function's container and removes its record from the database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "functions"
                ],
                "summary": "Remove a function",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Function ID",
                        "name": "functionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/functions/{functionID}/execute": {
            "post": {
                "description": "Sends a JSON payload to a function and returns the result.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "functions"
                ],
                "summary": "Execute a function",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Function ID",
                        "name": "functionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload for the function",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"result\": \"...\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "functions.Function": {
            "type": "object",
            "properties": {
                "container_id": {
                    "type": "string"
                },
                "container_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "function_name": {
                    "description": "The name of the function in the .py file",
                    "type": "string"
                },
                "handler_path": {
                    "description": "e.g., handler.handle",
                    "type": "string"
                },
                "host_port": {
                    "description": "The port on the host mapped to the container",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "description": "e.g., \"creating\", \"running\", \"stopped\", \"error\"",
                    "type": "string"
                }
            }
        }
    }
}